// renderer/models/transport-info.ts
export interface TransportInfo {
  itdOdvAssignedStops: {
    distance: string
    distanceTime: string
    gid: string
    isTransferStop: string
    mapName: string
    name: string
    nameWithPlace: string
    place: string
    stopID: string
    value: string
    vm: string
    x: string
    y: string
  }[]
}